fix(keta_otel): batch span export behind a bounded queue
- OtlpExporter no longer POSTs once per request: spans queue (bounded,
  default 2048) and a periodic sender ships batches (default 512 / 5s,
  BatchSpanProcessor-shaped defaults); a slow collector can no longer
  accumulate unbounded in-flight sockets
- overflow drops the oldest span and the count is reported in band on
  the next successful export, folded forward across failed ones — the
  same discipline as keta's log backlog
- failure/drop reporting moved from the per-request log to the
  exporter's own onWarn hook: a batch spans many requests, so a failure
  is not attributable to any one Context
- flush() still drains to quiescence (including spans enqueued
  mid-flush); close() cancels the timer so nothing pins the isolate;
  the per-POST timeout/socket-abort protection is unchanged
- durations record fractional milliseconds: elapsedMilliseconds
  truncated sub-ms handlers to 0, systematically undercounting the sum

+++ packages/keta_otel/lib/src/metrics.dart
@@ -13,13 +13,17 @@ library;
 /// attacker-controlled method or route — either grows memory without bound.
 class MetricsRegistry {
   final Map<_Key, int> _count = {};
-  final Map<_Key, int> _durationMsSum = {};
+  // `double`, not `int`: most in-process handlers finish in well under a
+  // millisecond, and truncating each sample to whole milliseconds before
+  // summing systematically undercounts (often to exactly 0) rather than
+  // just losing sub-ms precision on the total.
+  final Map<_Key, double> _durationMsSum = {};
 
   void record({
     required String method,
     required String route,
     required int status,
-    required int durationMs,
+    required double durationMs,
   }) {
     final key = _Key(method, route, status);
     _count.update(key, (v) => v + 1, ifAbsent: () => 1);
